ENG-8306 - cluster restart code starting nodes individually
Xcalar nodes need to be started simultaneously if their status
is going to be checked afterward.  Going back to xcalarctl, the
code used to start Xcalar has attempted to get the backend version
using xccli.  That attempt will not succeed until all usrnodes are
up and running.

The restartXcalar function was restarting the cluster node by node
and then checking the status of each start.  The only reason this
worked at all with the modern xcalar systemd unit was because the
status test was merely checking if usrnode was running or not.

This change causes the xcalar start to be run on all nodes at the
same time and then verify the status of each.

Change-Id: Ia2c7f7ad77de661af1d68a76278d7f881fb8aa54