ENG-8154 - make dcc support guardrails and gdb macros
This change adds support for guardrails to dcc.  It adds commands
to docker.yaml (and docker.json) to download the desired version
of Xcalar Guardrails (currently 0.0.1-1) from repo.xcalar.net
and install it within the docker-el7-base image.

Once that build image exists, dcc can be used to build cluster
docker containers that run with guardrails enabled with the
--guardrails flag:

dcc run -t xcalar/xcalar:2.3.0-10072 -n 1 --guardrails

This puts the variable XCE_USE_GUARDRAILS=1 in the
/etc/default/xcalar config file in each container, which
can be picked up by systemd scripts that understand the
variable.

This change also copies the Xcalar gdb macros from an
available branch of the XCE source tree (usually trunk
for Jenkins-generated VMs, the local branch in use on
other systems) during docker-xcalar-el7 build.  That
may not be ideal but it has more to with how Jenkins
passes around branch names.

When the --local flag is used, gdb macros from the
local XLRDIR are mapped into the containers.

Change-Id: Idfb17dcbdfd710533e986029488ffccb77c5ce2d