ENG-8765 - Creating meminfo file in dcc launch script
If "--maxswap" and "--maxmem" params are provided, we use this to
populate the meminfo file. If not, then the appropriate memory and swap
values are set to (current_value / num_nodes).

This meminfo file is then mapped into the containerized cluster, and
the usrnodes are notified to use the provided meminfo file through
the MEMINFO and CLUSTER_SAME_PHYSICAL_NODE env variables.

**Note, there are some stylistic changes as a result of running
shfmt_w.sh.

Change-Id: Ia4276dec3417a4ffb88fedc3d4ebdf9ed7350ed2