ENG-8959 - keep job info fresh
Keep the current list of currently active Jenkins jobs in the DB.
Will want to refer to that list from job-related dashboards so
as not to clutter them with stale information.

Also, arrange to update the job statistics at least every 24 hours
whether or not there have been recent builds.  Otherwise the info
on the Jenkins Job Stats dashboard is stale and misleading.

Change-Id: I769b42840d2c1e8c84549ce187ff2679e9d586d6