ENG-9006 Allow xcalar user local ansible access
Add an ansible inventory where the local host is
configured as an ansible "localclient" to avoid
using SSH unless needed.

In addition to the above, a random SSH keypair
is provisioned on node0 and distributed to the
cluster for the xcalar user on first boot. It
is not active by default. The admin user will
have to activate the key first, before xcalar
can ssh around the cluster. This keeps the security
high, as you still can't just ssh as xcalar by
default. Operationally this is significantly easier
as the admin only has to issue one command to
grant the xcalar user read access to the private
key on a single node.

Change-Id: I797b6bc8c58431a80684884f6b917c845cb77448