ENG-9294 add option to keep gce cluster up and skip cleanup
SqlScaleTest uses GCE to run and in some cases, to aid debugging, we
wanna keep the cluster up after a job has terminated (normally automatic
cleanup routine would teardown the cluster)

In order to do so the jenkins job has a new parameter KEEP_CLUSTER that
if set to 1 will cause the onExit function in the test to preserve the
GCE cluster. It's the user's reponsibility to clean up afterwards.

Change-Id: I229a8a2969f6c827f71ffd4d272e3dbf9b0aaee3