ENG-10038 - log parse fixes
Log parse errors should be noted in the log but not fatal
to processing.  We'll fix and reparse manually for a while.

Don't know how core analysis results were getting into mongo
with dots in the keys??? but it blew up in test, so that's
fixed here (and will need reparse).

Change-Id: I131408017b11e5391e6ebbdefbba17c8a5a88781