ENG-10686 - fix EfsAccessPoint mount behavior
user-data.sh is always creating and mounting the cluster directory
/cluster/${cluster_name}, but newer templates use an access point
that automatically sets the root directory path to
/cluster/${cluster_name}.   This effectively means the access point
mount is using /cluster/${cluster_name}/cluster/${cluster_name}
after upgrade.

This change causes user-data.sh to assume that the root directory
is correct when an accesspoint is specified in the mount options,
eliminating the double path pattern.

Change-Id: I48fc248c89b96a5023840ba7a73a84e436222b6d